+++ src/main/java/module-info.java
@@ -3,6 +3,6 @@ module org.example.project2 {
     requires javafx.fxml;
 
 
-    opens org.example.project2 to javafx.fxml;
-    exports org.example.project2;
+    opens temp to javafx.fxml;
+    exports temp;
 }
