--- src/main/java/org/example/project2/app.java
@@ -1,4 +1,0 @@
-package main.java.org.example.project2;
-
-public class app {
-}
