+++ src/main/java/temp/HelloApplication.java
@@ -1,4 +1,4 @@
-package org.example.project2;
+package temp;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
