+++ src/main/java/temp/app.java
@@ -1,0 +1,4 @@
+package temp;
+
+public class app {
+}
